+++ lib/solaar/ui/window.py
@@ -18,6 +18,8 @@
 
 import logging
 
+import gi
+
 from gi.repository import Gdk, GLib, Gtk
 from gi.repository.GObject import TYPE_PYOBJECT
 from logitech_receiver import hidpp10 as _hidpp10
@@ -26,14 +28,15 @@ from logitech_receiver.common import NamedInts as _NamedInts
 from logitech_receiver.status import KEYS as _K
 from solaar import NAME
 from solaar.i18n import _, ngettext
-# from solaar import __version__ as VERSION
-from solaar.ui import ui_async as _ui_async
 
 from . import action as _action
 from . import config_panel as _config_panel
 from . import icons as _icons
 from .about import show_window as _show_about_window
+from .common import ui_async as _ui_async
 from .diversion_rules import show_window as _show_diversion_window
+
+# from solaar import __version__ as VERSION
 
 logger = logging.getLogger(__name__)
 
@@ -47,7 +50,6 @@ _TREE_ICON_SIZE = Gtk.IconSize.BUTTON
 _INFO_ICON_SIZE = Gtk.IconSize.LARGE_TOOLBAR
 _DEVICE_ICON_SIZE = Gtk.IconSize.DND
 try:
-    import gi
     gi.check_version('3.7.4')
     _CAN_SET_ROW_NONE = None
 except (ValueError, AttributeError):
